+++ packages/plugin-image/web/src/image-viewer.jsx
@@ -262,11 +262,10 @@ class ImageViewer extends React.Component {
     } else if (this.hasAnchor()) {
       e.preventDefault();
       this.scrollToAnchor();
+    } else {
+      //!does expand mode needs to be available on on image click
+      !this.props.componentData.config.disableExpand && this.handleExpand(e);
     }
-    //!does expand mode needs to be available on on image click
-    // else {
-    //   this.handleExpand(e);
-    // }
   };
 
   handleRef = e => {
@@ -291,7 +290,7 @@ class ImageViewer extends React.Component {
     const { fallbackImageSrc, ssrDone } = this.state;
     const data = componentData || DEFAULTS;
     const { metadata = {} } = componentData;
-    const { disableExpand } = componentData.config; //! settings
+    const { disableExpand } = componentData.config; //! settings.config was undefined
 
     const hasExpand = !disableExpand && settings.onExpand;